@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5)
 project(graph)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_BUILD_TYPE "Debug" CACHE
       STRING "Choose the type of build." FORCE)
 
